set the jest test for backend
