+++ backend/controller/siteController.js
@@ -1,0 +1,4 @@
+import Site from '../models/siteModel.js'
+// asyncHandler is used to simplify the syntax of the promise catch error handling
+import asyncHandler from 'express-async-handler'
+
