+++ backend/routes/siteRoutes.js
@@ -1,0 +1,7 @@
+import express from 'express'
+
+const router = express.Router()
+
+router.route('/').get(getSites)
+
+export default router
