+++ package.json
@@ -12,9 +12,9 @@
     "server": "nodemon backend/server",
     "client": "yarn --cwd frontend/ start",
     "dev": "yarn server & yarn client",
+    "test": "jest",
     "data:import": "node backend/seeder",
-    "data:destroy": "node backend/seeder -d",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "data:destroy": "node backend/seeder -d"
   },
   "repository": {
     "type": "git",
@@ -37,6 +37,11 @@
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^6.7.0",
     "multer": "^1.4.5-lts.1",
-    "nodemon": "^2.0.20"
+    "nodemon": "^2.0.20",
+    "supertest": "^6.3.1"
+  },
+  "devDependencies": {
+    "@babel/plugin-transform-modules-commonjs": "^7.19.6",
+    "jest": "^29.2.2"
   }
 }
